Model/main_service/pyproject.toml
[project]
name = "main-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aioredis>=2.0.1",
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "fastapi-cache2>=0.2.2",
    "fastapi-users[sqlalchemy]>=14.0.1",
    "fastapi[standard]>=0.115.12",
    "opencv-python>=4.11.0.86",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.9.1",
    "redis>=5.2.1",
    "sqlalchemy[asyncio]>=2.0.40",
]
